+++ src/database.py
@@ -1,6 +1,7 @@
 from enum import Enum
 import re
 import distutils.util
+import tempora
 
 from . import settings
 
@@ -15,68 +16,70 @@ class DatabaseType(Enum):
 
 KNOWN_IMAGES = {
     # MySQL
-    "mysql": DatabaseType.mysql,
-    "bitnami/mysql": DatabaseType.mysql,
-    "mysql/mysql-server": DatabaseType.mysql,
+    "mysql": "mysql",
+    "bitnami/mysql": "mysql",
+    "mysql/mysql-server": "mysql",
     # MariaDB
-    "mariadb": DatabaseType.mariadb,
-    "bitnami/mariadb": DatabaseType.mariadb,
-    "linuxserver/mariadb": DatabaseType.mariadb,
-    "mariadb/server": DatabaseType.mariadb,
+    "mariadb": "mariadb",
+    "bitnami/mariadb": "mariadb",
+    "linuxserver/mariadb": "mariadb",
+    "mariadb/server": "mariadb",
     # Postgres
-    "postgres": DatabaseType.postgres,
-    "bitnami/postgresql": DatabaseType.postgres,
+    "postgres": "postgres",
+    "bitnami/postgresql": "postgres",
 }
 
 TYPE_DEFAULTS = {
-    DatabaseType.unknown: {
-        "port": 0,
+    "unknown": {
+        "port": "0"
     },
-    DatabaseType.mysql: {
-        "port": 3306,
+    "mysql": {
+        "port": "3306",
     },
-    DatabaseType.mariadb: {
-        "port": 3306,
+    "mariadb": {
+        "port": "3306",
     },
-    DatabaseType.postgres: {
-        "port": 5432,
+    "postgres": {
+        "port": "5432",
+    }
+}
+
+RETENTION_DEFAULTS = {
+    "none": {
+        "dump_timestamp": "false",
+        "retention_min_count": "1",
+        "retention_min_age": "0s",
+        "retention_max_count": "1",
+        "retention_max_age": "0s"
+    },
+    "simple": {
+        "dump_timestamp": "true",
+        "retention_min_count": "10",
+        "retention_min_age": "0s",
+        "retention_max_count": "0",
+        "retention_max_age": "1 month"
+    },
+    "all": {
+        "dump_timestamp": "true",
+        "retention_min_count": "1",
+        "retention_min_age": "0s",
+        "retention_max_count": "0",
+        "retention_max_age": "0s"
     }
 }
 
 
 class Database:
-    IMAGE_REGEX = re.compile("^(.+?)(?::.+)?$")
-
     def __init__(self, container, global_labels):
         self._load_labels(global_labels)
         local_labels = self._get_labels_from_container(container)
         self._load_labels(local_labels)
         self._resolve_labels(container)
 
-    def _load_labels(self, values):
-        if "enable" in values:
-            self.enable = values["enable"]
-        if "type" in values:
-            self.type = values["type"]
-        if "port" in values:
-            self.port = values["port"]
-        if "username" in values:
-            self.username = values["username"]
-        if "password" in values:
-            self.password = values["password"]
-        if "dump_name" in values:
-            self.dump_name = values["dump_name"]
-        if "dump_timestamp" in values:
-            self.dump_timestamp = distutils.util.strtobool(
-                values["dump_timestamp"])
-        if "compress" in values:
-            self.compress = distutils.util.strtobool(values["compress"])
-        if "compression_level" in values:
-            self.compression_level = int(values["compression_level"])
-        if "encrypt" in values:
-            self.encrypt = distutils.util.strtobool(values["encrypt"])
-        if "encryption_key" in values:
-            self.encryption_key = values["encryption_key"]
+    def _load_labels(self, values, only_auto_values=False):
+        for key in settings.LABEL_DEFAULTS.keys():
+            if key in values and (not only_auto_values or getattr(self, key) == "auto"):
+                setattr(self, key, values[key])
 
     def _get_labels_from_container(self, container):
         labels = {}
@@ -89,25 +92,33 @@ class Database:
         return labels
 
     def _resolve_labels(self, container):
-        try:
-            self.type = DatabaseType[self.type]
-        except AttributeError:
-            self.type = DatabaseType.unknown
-
-        if self.type == DatabaseType.auto:
+        # Resolve database type
+        if self.type == "auto":
+            image_regex = re.compile("^(.+?)(?::.+)?$")
             for tag in container.image.tags:
-                matches = self.IMAGE_REGEX.match(tag)
-                search_part = matches.group(1)
-                matched_type = KNOWN_IMAGES.get(search_part, None)
+                matches = image_regex.match(tag)
+                matched_type = KNOWN_IMAGES.get(matches.group(1), None)
                 if matched_type is not None:
                     self.type = matched_type
                     break
-            if self.type == DatabaseType.auto:
-                self.type = DatabaseType.unknown
+            if self.type == "auto":
+                self.type = "unknown"
 
-        defaults = TYPE_DEFAULTS.get(self.type)
+        # Resolve default values based on other attributes
+        self._load_labels(TYPE_DEFAULTS.get(self.type, {}), True)
+        self._load_labels(RETENTION_DEFAULTS.get(
+            self.retention_policy, {}), True)
 
-        if self.port == "auto":
-            self.port = defaults.get("port")
-        else:
-            self.port = int(self.port)
+        # Other
+        self.type = DatabaseType[self.type]
+        self.port = int(self.port)
+        self.compress = distutils.util.strtobool(self.compress)
+        self.compression_level = int(self.compression_level)
+        self.encrypt = distutils.util.strtobool(self.encrypt)
+        self.dump_timestamp = distutils.util.strtobool(self.dump_timestamp)
+        self.retention_min_count = max(int(self.retention_min_count), 1)
+        self.retention_min_age = tempora.parse_timedelta(
+            self.retention_min_age)
+        self.retention_max_count = max(int(self.retention_max_count), 0)
+        self.retention_max_age = tempora.parse_timedelta(
+            self.retention_max_age)